+++ .opspec/.common/go-build-entrypoint.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-
-go get -t ./... && \
-go build -a -o .tmp/opctl-${GOOS}-x86_64${OUTPUT_FILE_EXTENSION}