+++ .opspec/publish-coverage-to-codecov/entrypoint.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-
-cat cli.coverprofile > coverage.txt && \
-curl -s https://codecov.io/bash | bash -s