+++ .opspec/test/entrypoint.sh
@@ -1,5 +1,0 @@
-#!/bin/sh -e
-
-go get -t ./...
-go get github.com/onsi/ginkgo/ginkgo
-/golang/bin/ginkgo -r -cover