+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Added
+- collection subcommand
+- set collection description use case
+### Fixed
+- engine API errors not handled
+
 ## 0.1.0 - 2016-05-27